History screen UI

+++ app/src/main/java/uk/ac/tees/mad/bpmtracker/database/Repository.kt
@@ -3,7 +3,7 @@ package uk.ac.tees.mad.bpmtracker.database
 import kotlinx.coroutines.flow.Flow
 
 interface Repository {
-    suspend fun addRecord(entity: RecordEntity):Int
+    suspend fun addRecord(entity: RecordEntity):Long
     fun getAllRecords(): Flow<List<RecordEntity>>
     suspend fun deleteRecord(entity: RecordEntity)
 }
