+++ app/src/main/java/uk/ac/tees/mad/bpmtracker/database/RepositoryImpl.kt
@@ -6,8 +6,8 @@ import javax.inject.Inject
 class RepositoryImpl @Inject constructor(
     private val dao: RecordDao
 ):Repository {
-    override suspend fun addRecord(entity: RecordEntity): Int {
-        return dao.addRecord(entity).toInt()
+    override suspend fun addRecord(entity: RecordEntity):Long {
+        return dao.addRecord(entity)
     }
 
     override fun getAllRecords(): Flow<List<RecordEntity>> {
